new jpress

+++ jpress-model/src/main/java/io/jpress/model/base/BaseUtm.java
@@ -17,12 +17,12 @@ public abstract class BaseUtm<M extends BaseUtm<M>> extends JbootModel<M> implem
 		return getStr("id");
 	}
 
-	public void setUserId(java.lang.Integer userId) {
+	public void setUserId(java.lang.Long userId) {
 		set("user_id", userId);
 	}
 	
-	public java.lang.Integer getUserId() {
-		return getInt("user_id");
+	public java.lang.Long getUserId() {
+		return getLong("user_id");
 	}
 
 	public void setAnonym(java.lang.String anonym) {
